introduce lzfw_getattrat
Gets the value of an extended attribute on an object known
by its vnode handle.

The interface is made more similar to the linux/POSIX xattr
interface (than the current lzfw_getattr):

* the data is placed in a pre-allocated buffer (not malloc'd)
* the bytes written are placed in a size argument (pointer)
* the mechanism to read xattr stored size is implemented

It's possible that lzfw_getattr and lzfw_setattr will change at
some point to be more similar to this interface.  It's also
likely that a subfile interface to xattrs will be exposed.

Signed-off-by: Matt Benjamin <matt@cohortfs.com>