Expose vnode.h and dependencies.
This brings a large number of Solaris includes into portable--
possibly all except spa.h and dependencies.

The only known defect for internal compilation of libzfswrap is
failure to include Linux asm/stat.h such that struct stat64 is
defined--it should be controlled by __USE_LARGEFILE64 (which is
defined) but for some reason, this is not working.

A known limitation is that _KERNEL is not defined by the include
path, which may need to be changed.

Signed-off-by: Matt Benjamin <matt@cohortfs.com>